+++ src/pages/home.js
@@ -1,5 +1,5 @@
-import { Header } from "./header";
-import { Footer } from "./footer";
+import { Header } from "../components/header";
+import { Footer } from "../components/footer";
 
 export const HomePage = () => {
   const header = Header();
